build,ci: Add build option to pin `RuntimeFrameworkVersion` and use it for all pre-built self-contained artifacts
Before this, half of the artifacts were built with the latest version of dotnet from actions/setup-dotnet, and another half with a pinned version corresponding to the used dotnet SDK container image.
Now, all artifacts will be built with the same pinned version (the alternative of using non-pinned versions of container images breaks reproducibility and lacks the clarity of what version was/is used).