ci: Use Trivy scan-type=image for vuln-library on commit events
Deps in `packages.lock.json` don't include dotnet runtime meta-packages, which are the only real deps in this project and are the main source of vulnerabilities.
Also, all non-runtime packages are already checked for vulnerabilities by dotnet during build (including transitive since NuGetAuditMode=all by default for net10.0), so scanning it with Trivy is redundant.