build(docker): Add apk upgrade to app-base images
Considering that the base images aren't pinned, and that they are updated regularly, only a few packages will be updated or none at all.
Layer-size-wise this should be negligent, so it's worth to trade for the timely non-Critical security updates
(dotnet upstream images are updated immediately only for Critical CVEs, otherwise they are mostly updated monthly).
This is generally the best practice even if upstream images are updated regularly (see https://github.com/docker/docs/pull/12571).